# src/main/java/com/bank/accounts/application/api/ServicePointsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.accounts.application.api;

import com.bank.accounts.application.schemas.request.ServicePointRequest;
import com.bank.accounts.application.schemas.response.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-18T07:29:52.914264900-05:00[America/Lima]")
@Validated
@Tag(name = "service-points", description = "the service-points API")
public interface ServicePointsApi {

    /**
     * POST /service-points/owners/{accountId} : Crear puntos de servicio
     *
     * @param accountId           (required)
     * @param servicePointRequest (required)
     * @return (status code 200)
     * or  (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "createServicePoint",
            summary = "Crear puntos de servicio",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TrackingResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/service-points/owners/{accountId}",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    Mono<TrackingResponse> createServicePoint(
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @Parameter(name = "ServicePointRequest", description = "", required = true) @Valid @RequestBody Mono<ServicePointRequest> servicePointRequest,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );


    /**
     * DELETE /service-points/{servicePointId}/owners/{accountId} : Borrar puntos de servicio
     *
     * @param servicePointId (required)
     * @param accountId      (required)
     * @return (status code 200)
     * or  (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "deleteServicePoint",
            summary = "Borrar puntos de servicio",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TrackingResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/service-points/{servicePointId}/owners/{accountId}",
            produces = {"application/json"}
    )
    Mono<TrackingResponse> deleteServicePoint(
            @Parameter(name = "servicePointId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("servicePointId") Long servicePointId,
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );


    /**
     * PUT /service-points/{servicePointId}/owners/{accountId} : Actualizar puntos de servicio
     *
     * @param servicePointId      (required)
     * @param accountId           (required)
     * @param servicePointRequest (required)
     * @return (status code 200)
     * or  (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "editServicePoint",
            summary = "Actualizar puntos de servicio",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TrackingResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/service-points/{servicePointId}/owners/{accountId}",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    Mono<TrackingResponse> editServicePoint(
            @Parameter(name = "servicePointId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("servicePointId") Long servicePointId,
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @Parameter(name = "ServicePointRequest", description = "", required = true) @Valid @RequestBody Mono<ServicePointRequest> servicePointRequest,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );


    /**
     * GET /service-points/{servicePointId} : Obtener puntos de servicio
     *
     * @param servicePointId (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "obtainServicePoint",
            summary = "Obtener puntos de servicio",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ServicePointResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/service-points/{servicePointId}",
            produces = {"application/json"}
    )
    Mono<ServicePointResponse> obtainServicePoint(
            @Parameter(name = "servicePointId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("servicePointId") Long servicePointId,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );


    /**
     * GET /service-points/{servicePointId}/owners/{accountId} : Obtener puntos de servicio de un propietario
     *
     * @param servicePointId (required)
     * @param accountId      (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "obtainServicePointOwner",
            summary = "Obtener puntos de servicio de un propietario",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ServicePointResponse.class)))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/service-points/{servicePointId}/owners/{accountId}",
            produces = {"application/json"}
    )
    Flux<ServicePointResponse> obtainServicePointOwner(
            @Parameter(name = "servicePointId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("servicePointId") Long servicePointId,
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );

}