# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: BANK-API-V1
  description: ''
  version: 1.0.0
...
paths:
...
  /service-points/owners/{accountId}:
    post:
      summary: Crear puntos de servicio
      deprecated: false
      description: ''
      operationId: createServicePoint
      tags: []
      parameters:
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePointRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
        '202':
          $ref: '#/components/responses/Tracking'
          description: ''
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /service-points/{servicePointId}/owners/{accountId}:
    put:
      summary: Actualizar puntos de servicio
      deprecated: false
      description: ''
      operationId: editServicePoint
      tags: []
      parameters:
        - name: servicePointId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePointRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
        '202':
          $ref: '#/components/responses/Tracking'
          description: ''
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
    delete:
      summary: Borrar puntos de servicio
      deprecated: false
      description: ''
      operationId: deleteServicePoint
      tags: []
      parameters:
        - name: servicePointId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
        '202':
          $ref: '#/components/responses/Tracking'
          description: ''
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
    get:
      summary: Obtener puntos de servicio de un propietario
      deprecated: false
      description: ''
      operationId: obtainServicePointOwner
      tags: []
      parameters:
        - name: servicePointId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicePointResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /service-points/{servicePointId}:
    get:
      summary: Obtener puntos de servicio
      deprecated: false
      description: ''
      operationId: obtainServicePoint
      tags: []
      parameters:
        - name: servicePointId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePointResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
...
components:
  schemas:
...
    CardResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        numberCard:
          type: string
          pattern: ^\d{13,19}$
        cvc:
          type: string
          pattern: ^\d{3,4}$
        date:
          type: string
          pattern: ^(0[1-9]|1[0-2])\/\d{2}$
        accountId:
          type: integer
          format: int64
        cardType:
          type: string
          pattern: ^(CREDIT|DEBIT)$
        creditDebitAccount:
          anyOf:
            - $ref: '#/components/schemas/DebitAccountResponse'
            - $ref: '#/components/schemas/CreditAccountResponse'
      required:
        - numberCard
        - cvc
        - date
        - accountId
        - cardType
        - cardId
        - creditDebitAccount
    DebitAccountResponse:
      type: object
      properties:
        debitAccountId:
          type: integer
          format: int64
        cardId:
          type: string
        amount:
          type: number
        currency:
          type: string
          pattern: ^(PEN|USD)$
        debitAccountType:
          type: string
          pattern: ^(AHORROS|CORRIENTE)$
      required:
        - amount
        - currency
        - debitAccountType
        - cardId
        - debitAccountId
    CreditAccountResponse:
      type: object
      properties:
        creditAccountId:
          type: integer
          format: int64
        amount:
          type: number
        maxAccount:
          type: number
        facturationDate:
          type: string
          format: date
        lastPaymentDate:
          type: string
          format: date
        currency:
          type: string
          pattern: ^(PEN|USD)$
      required:
        - amount
        - maxAccount
        - facturationDate
        - lastPaymentDate
        - currency
        - creditAccountId
...
    AccountResponse:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        documentType:
          type: string
          pattern: ^(DNI|CE|PS|RUC)$
        documentNumber:
          type: string
        phone:
          type: string
          pattern: ^9\d{8}$
        email:
          type: string
          format: email
        names:
          type: string
        lastNames:
          type: string
        dateOfBirth:
          type: string
          format: date
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
      required:
        - documentType
        - documentNumber
        - phone
        - email
        - names
        - lastNames
        - dateOfBirth
        - accountId
        - cards
...
    ServicePointRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ServicePointResponse:
      type: object
      properties:
        servicePointId:
          type: integer
          format: int64
        name:
          type: string
        accountOwner:
          $ref: '#/components/schemas/AccountResponse'
      required:
        - name
        - accountOwner
        - servicePointId
    TrackingResponse:
      type: object
      properties:
        operationDate:
          type: string
          format: date-time
        trackingId:
          type: string
          format: uuid
      required:
        - operationDate
        - trackingId
    BadRequestResponse:
      type: object
      properties:
        description:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - description
        - errors
    InternalErrorResponse: &ref_0
      type: object
      properties:
        description:
          type: string
      required:
        - description
...
  responses:
    Tracking:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TrackingResponse'
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundResponse'
    BadRequest:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestResponse'
    InternalError:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  securitySchemes: {}
servers: []
security: []